refactor: more s3 functions to upload service
