+++ backend/src/email/middlewares/email-template.middleware.ts
@@ -144,7 +144,7 @@ const uploadCompleteHandler = async (
 
     // extract s3Key from transactionId
     const { transaction_id: transactionId, filename } = req.body
-    const s3Key = UploadService.extractS3Key(transactionId)
+    const s3Key = UploadService.extractParamsFromJwt(transactionId) as string
 
     // check if template exists
     const emailTemplate = await EmailTemplateService.getFilledTemplate(
