+++ backend/src/sms/middlewares/sms-template.middleware.ts
@@ -135,7 +135,7 @@ const uploadCompleteHandler = async (
 
     // extract s3Key from transactionId
     const { transaction_id: transactionId, filename } = req.body
-    const s3Key = UploadService.extractS3Key(transactionId)
+    const s3Key = UploadService.extractParamsFromJwt(transactionId) as string
 
     // check if template exists
     const smsTemplate = await SmsTemplateService.getFilledTemplate(+campaignId)
